+++ deadcode/deadcode.py
@@ -1,0 +1,5 @@
+def deadcode():
+    """
+    don't call this function from anywhere
+    intentional dead code for testing purposes
+    """