+++ tests/test_config.py
@@ -27,6 +27,7 @@ def test_cli_args():
         ignore_names=["name1", "name2"],
         make_whitelist=True,
         min_confidence=10,
+        format="relative",
         sort_by_size=True,
         verbose=True,
     )
@@ -39,6 +40,7 @@ def test_cli_args():
             "--min-confidence=10",
             "--sort-by-size",
             "--verbose",
+            "--format=relative",
             "path1",
             "path2",
         ]
@@ -60,6 +62,7 @@ def test_toml_config():
         min_confidence=10,
         sort_by_size=True,
         verbose=True,
+        format="relative",
     )
     data = StringIO(
         dedent(
@@ -73,6 +76,7 @@ def test_toml_config():
         sort_by_size = true
         verbose = true
         paths = ["path1", "path2"]
+        format = "relative"
         """
         )
     )
@@ -97,6 +101,7 @@ def test_config_merging():
         min_confidence = 10
         sort_by_size = false
         verbose = false
+        format = "relative"
         paths = ["toml_path"]
         """
         )
@@ -108,6 +113,7 @@ def test_config_merging():
         "--make-whitelist",
         "--min-confidence=20",
         "--sort-by-size",
+        "--format=relative",
         "--verbose",
         "cli_path",
     ]
@@ -120,6 +126,7 @@ def test_config_merging():
         make_whitelist=True,
         min_confidence=20,
         sort_by_size=True,
+        format="relative",
         verbose=True,
     )
     assert result == expected