+++ vulture/utils.py
@@ -1,7 +1,15 @@
+"""module: utils"""
+
+# standard imports
 import ast
 import os
 import sys
 import tokenize
+from pathlib import Path
+
+from .config import RELATIVE_PATH_FORMAT, ABSOLUTE_PATH_FORMAT
+
+EMPTY_PATH = ""
 
 
 class VultureInputException(Exception):
@@ -26,15 +34,14 @@ def _safe_eval(node, default):
         results = [_safe_eval(value, default) for value in node.values]
         if isinstance(node.op, ast.And):
             return all(results)
-        else:
-            return any(results)
-    elif isinstance(node, ast.UnaryOp) and isinstance(node.op, ast.Not):
+        return any(results)
+    if isinstance(node, ast.UnaryOp) and isinstance(node.op, ast.Not):
         return not _safe_eval(node.operand, not default)
-    else:
-        try:
-            return ast.literal_eval(node)
-        except ValueError:
-            return default
+
+    try:
+        return ast.literal_eval(node)
+    except ValueError:
+        return default
 
 
 def condition_is_always_false(condition):
@@ -45,12 +52,36 @@ def condition_is_always_true(condition):
     return _safe_eval(condition, False)
 
 
-def format_path(path):
-    try:
-        return path.relative_to(os.curdir)
-    except ValueError:
-        # Path is not below the current directory.
-        return path
+def format_path(
+    filename_path_str: str,
+    format_str: str,
+    *args,
+    format_id: str = RELATIVE_PATH_FORMAT,
+) -> str:
+    if not filename_path_str:
+        return EMPTY_PATH
+    if format_id not in (RELATIVE_PATH_FORMAT, ABSOLUTE_PATH_FORMAT):
+        raise ValueError(f"path format {format_id} uknown.")
+    _path = Path(filename_path_str)
+    if format_id == RELATIVE_PATH_FORMAT:
+        _path_str = str(filename_path_str)
+        _relpath_str = os.path.relpath(_path_str, start=os.curdir)
+        _use_path_str = (
+            _path_str if _relpath_str.startswith("..") else _relpath_str
+        )
+        _formatted_path_str = (
+            format_str.format(_use_path_str, *args)
+            if format_str
+            else _use_path_str
+        )
+        return _formatted_path_str
+    if format_id == ABSOLUTE_PATH_FORMAT:
+        _abs_path = _path.resolve(strict=True)
+        if format_str:
+            return format_str.format(str(_abs_path), *args)
+        return str(_abs_path)
+    # should never get here
+    raise ValueError(f"path format {format_id} uknown.")
 
 
 def get_decorator_name(decorator):
@@ -93,14 +124,15 @@ def read_file(filename):
         with tokenize.open(filename) as f:
             return f.read()
     except (SyntaxError, UnicodeDecodeError) as err:
-        raise VultureInputException(err)
+        raise VultureInputException(err) from err
 
 
 class LoggingList(list):
     def __init__(self, typ, verbose):
         self.typ = typ
         self._verbose = verbose
-        return list.__init__(self)
+        # return list.__init__(self)
+        list.__init__(self)
 
     def append(self, item):
         if self._verbose:
@@ -112,7 +144,8 @@ class LoggingSet(set):
     def __init__(self, typ, verbose):
         self.typ = typ
         self._verbose = verbose
-        return set.__init__(self)
+        # return set.__init__(self)
+        set.__init__(self)
 
     def add(self, name):
         if self._verbose: